break test.js into small pieces
