+++ common/constants.js
@@ -1,0 +1,4 @@
+exports.TYPE_SERVICE_INSTANCE = 'service_instance'
+exports.TYPE_SERVICE_BINDING = 'service_binding'
+exports.ACTION_CREATE = 'create'
+exports.ACTION_UPDATE = 'update'
